Rework vpc service to be not version aware (#249)
Rework vpc service to be not version aware

there are 4 different url types in the VPC service. This is not
manageable. Therefore make the vpc proxy not supporting version
(everything is a version 1) and make resources instead contain version
path.

move vpcp implementation from v2 to v1
add vpc resource

Reviewed-by: None <None>
Reviewed-by: Artem Goncharov <Artem.goncharov@gmail.com>
Reviewed-by: Vladimir Vshivkov <None>
Reviewed-by: OpenTelekomCloud Bot <None>